TeamDHEINCHOO/SECUREGUARD_SOURCE_CODE/New_project/New project/integrations/final/requirements.txt
flask
requests
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langgraph
google-generativeai